# conductor-backend/src/main/resources/openapi-v2.yaml
openapi: "3.0.3"
info:
  title: Conductor API v2
  description: >-
    Conductor API v2 — the canonical Work Item resource. Supersedes the legacy
    v1 `issues` core endpoints (see docs/legacy-v1-deprecation.md). This spec is
    additive: v1 remains served unchanged.
  version: "2.0.0"
servers:
  - url: /api/v2
    description: API v2
...
paths:
  /projects/{projectId}/work-items:
    post:
      tags:
        - work-items
      summary: Create a Work Item in a project
      operationId: createWorkItem
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkItemRequest"
      responses:
        "201":
          description: Work Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItemResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found or caller is not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - work-items
      summary: List Work Items in a project
      operationId: listWorkItems
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          description: Filter by Work Item type (validated against the bound Workflow's types).
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter by Work Item status (a Workflow-defined status id, e.g. DRAFT).
          schema:
            type: string
        - name: workflow
          in: query
          required: false
          description: Filter by bound Workflow slug (e.g. ENGINEERING) — for per-Workflow view pages.
          schema:
            type: string
      responses:
        "200":
          description: List of Work Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkItemResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found or caller is not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/work-items/by-display/{displayId}:
    get:
      tags:
        - work-items
      summary: Get a Work Item by its human-readable display id
      description: >-
        Resolve a Work Item by its human-readable display id (e.g. "COND-42").
        The trailing number is the project-scoped sequence number.
      operationId: getWorkItemByDisplayId
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: displayId
          in: path
          required: true
          description: Human-readable Work Item id (e.g. "COND-42").
          schema:
            type: string
      responses:
        "200":
          description: Work Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItemResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Work Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/work-items/{workItemId}:
    get:
      tags:
        - work-items
      summary: Get Work Item detail
      operationId: getWorkItem
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: workItemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Work Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItemResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Work Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - work-items
      summary: Update a Work Item
      operationId: patchWorkItem
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: workItemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchWorkItemRequest"
      responses:
        "200":
          description: Work Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItemResponse"
        "400":
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Work Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - work-items
      summary: Delete a Work Item
      operationId: deleteWorkItem
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: workItemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Work Item deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Work Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/work-items/{workItemId}/available-transitions:
    get:
      tags:
        - work-items
      summary: List the valid next transitions for a Work Item (doer projection)
      description: >-
        Returns only the transitions the calling actor may take from the Work Item's
        current status, computed server-side from the active (PUBLISHED) Workflow
        definition. This is the doer projection — it never exposes the raw statechart,
        gate conditions, or bound Steps. A Review-gated transition does not appear here
        until its Review is satisfied.
      operationId: getWorkItemAvailableTransitions
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: workItemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The valid next transitions for this actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableTransitionsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Work Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key authentication (Bearer token)
  schemas:
    WorkItemResponse:
      type: object
      required:
        - id
        - projectId
        - type
        - title
        - status
        - createdBy
        - createdAt
        - updatedAt
        - sequenceNumber
        - displayId
      properties:
        id:
          type: string
        projectId:
          type: string
        type:
          type: string
          description: Work Item type — a Workflow-defined string (e.g. PRD).
        title:
          type: string
        description:
          type: string
        status:
          type: string
          description: Current status — a Workflow-defined status id (e.g. DRAFT).
        workflow:
          type: string
          description: The bound Workflow definition id (slug, e.g. ENGINEERING).
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sequenceNumber:
          type: integer
        displayId:
          type: string
          description: Human-readable ID (e.g. "COND-42")
        unresolvedCommentCount:
          type: integer
        assignee:
          $ref: "#/components/schemas/WorkItemAssignee"
    WorkItemAssignee:
      type: object
      required:
        - userId
        - name
      properties:
        userId:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
    CreateWorkItemRequest:
      type: object
      required:
        - workflow
        - type
        - title
      properties:
        workflow:
          type: string
          description: >-
            Required — the lifecycle Workflow definition id (slug, e.g. ENGINEERING) that governs this Work
            Item. The canonical v2 surface never silently defaults it; discover options via
            `GET /projects/{projectId}/workflows?lifecycle=true`. (The legacy v1 `/issues` create keeps an
            ENGINEERING default for back-compat.)
        type:
          type: string
          description: >-
            Work Item type — a Workflow-defined string (e.g. PRD), validated against the
            chosen Workflow's allowed types.
        title:
          type: string
          maxLength: 255
        description:
          type: string
    PatchWorkItemRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          description: >-
            Target status — a Workflow-defined status id (e.g. IN_REVIEW). The move is
            validated against the bound Workflow and the Review gate.
        assigneeId:
          type: string
          nullable: true
          description: User ID to assign to this Work Item, or null to unassign
    AvailableTransition:
      type: object
      required:
        - toStatus
        - label
      properties:
        toStatus:
          type: string
          description: The status this transition moves the Work Item to.
        label:
          type: string
          description: Human-readable action label (e.g. "Submit for review").
        requiresReview:
          type: boolean
          description: Whether taking this transition will open a Review gate.
    AvailableTransitionsResponse:
      type: object
      required:
        - workflow
        - currentStatus
        - transitions
      properties:
        workflow:
          type: string
          description: The Work Item's Workflow definition id (slug).
        currentStatus:
          type: string
        noun:
          type: string
          description: The Workflow's display noun (e.g. Issue, Post).
        transitions:
          type: array
          items:
            $ref: "#/components/schemas/AvailableTransition"
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string